REF_CODE_REFACTOR: hard_swish Refactoring the reference code for hard_swish operator.
BUG=refactoring existing code.
