+++ tensorflow/lite/micro/kernels/BUILD
@@ -143,6 +143,7 @@ cc_library(
         "gather.cc",
         "gather_nd.cc",
         "hard_swish.cc",
+        "hard_swish_common.cc",
         "if.cc",
         "l2_pool_2d.cc",
         "l2norm.cc",
@@ -188,6 +189,7 @@ cc_library(
         "depthwise_conv.h",
         "ethosu.h",
         "fully_connected.h",
+        "hard_swish.h",
         "micro_ops.h",
         "pooling.h",
         "quantize.h",
